authorization

+++ controllers/contactController.js
@@ -3,27 +3,33 @@ const Contact = require("../models/contactModel");
 
 // @desc Get all contacts
 // @route GET /api/contacts
-// @access public
+// @access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 })
 
 // @desc Get one contact
 // @route GET /api/contacts/:id
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User doesn't have permission to change others contacts")
+    }
+
     res.status(200).json(contact);
 })
 
 // @desc Create a contact
 // @route POST /api/contacts/:id
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
     const {name, email, phone} = req.body;
     if(!name || !email || !phone) {
@@ -34,7 +40,8 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     });
 
     res.status(201).json(contact);
@@ -42,7 +49,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 // @desc Update a contact
 // @route PUT /api/contacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
@@ -61,12 +68,17 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc Delete a contact
 // @route DELETE /api/contacts/:id
-// @access public
+// @access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id); 
     if(!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User doesn't have permission to delete others contacts")
     }
 
     await Contact.deleteOne(contact);
